refactor(auth): rename auth client
